calculo venda

+++ pdv-integrado/src/main/java/br/unipar/pdvintegrado/models/Cliente.java
@@ -11,7 +11,7 @@ public class Cliente {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @NotBlank
     @Length(min = 10, max = 180)
@@ -28,18 +28,18 @@ public class Cliente {
     public Cliente() {
     }
 
-    public Cliente(int id, String nome, String telefone, String email) {
+    public Cliente(long id, String nome, String telefone, String email) {
         this.id = id;
         this.nome = nome;
         this.telefone = telefone;
         this.email = email;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
